requirements.txt
torch>=1.10
numpy>=1.21
scipy>=1.7
scikit-learn>=1.0
PyYAML>=6.0
tqdm>=4.64
torch-geometric>=2.0
